// src/app/shared/models/MediumCharacteristic.model.ts
export interface MediumCharacteristic {
	id: string,
	city: string;
	contactType: string;
	country: string;
	emailAddress: string;
	faxNumber: string;
	phoneNumber: string;
	postCode: string;
	socialNetworkId: string;
	stateOrProvince: string;
	street1: string;
	street2: string;
	baseType: string;
	schemaLocation: string;
	type: string;
}